pyproject.toml
[project]
name = "domain-specific-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "numpy>=1.25.0",
    "pypdf2>=3.0.1",
    "python-docx>=0.8.11",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.25.0",
]
